requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anthropic==0.40.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.4.0
Backtesting==0.3.3
beautifulsoup4==4.14.2
bitarray==3.7.2
bokeh==3.8.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
contourpy==1.3.2
cytoolz==1.1.0
decorator==5.2.1
distro==1.9.0
elevenlabs>=2.15.0
eth-account==0.11.0
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.3.0
executing==2.2.1
ffmpeg-python==0.2.0
filelock==3.20.0
filetype==1.2.0
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.185.0
google-auth==2.42.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.71.0
groq==0.16.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==0.3.1
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperliquid-python-sdk==0.20.0
idna==3.11
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonalias==0.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.44.0
lru-dict==1.2.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multitasking==0.0.12
narwhals==2.10.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai==1.59.5
openai-whisper==20231117
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
#pandas-ta-classic==0.3.36
parsimonious==0.10.0
parso==0.8.5
pathlib==1.0.1
peewee==3.18.2
pillow==10.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
pure_eval==0.2.3
py-ecc==8.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.12.0
Pygments==2.19.2
pymssql==2.3.8
pyotp==2.9.0
pyparsing==3.2.5
PyPDF2==3.0.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.31.0
rlp==4.1.0
rpds-py==0.28.0
rsa==4.9.1
schedule==1.2.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
solana==0.35.1
solders==0.21.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
TA-Lib @ file:///D:/dev/moon-dev-ai-agents/ta_lib-0.6.8-cp312-cp312-win_amd64.whl
termcolor==2.3.0
tiktoken==0.12.0
tomli==2.3.0
toolz==1.1.0
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
twikit==2.2.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
web3==6.20.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==11.0.3
whisper==1.1.10
xyzservices==2025.4.0
yarl==1.22.0
yfinance==0.2.43
youtube-transcript-api==0.6.2
watchdog>=3.0.0
mypy-extensions>=1.0.0
tldextract>=5.0.0
greenlet>=1.0.0
click>=7.0
